# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Engage API",
    "version": "1.0.0",
    "description": "選挙活動管理プラットフォーム Engage のAPIドキュメント"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/api/leads/{id}/notifications": {
      "post": {
        "summary": "リードの通知設定",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "days": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードの通知設定",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationPreference"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{id}": {
      "get": {
        "summary": "リード詳細取得",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リード詳細",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lead"
                }
              }
            }
          }
        }
      }
    },
    "/api/leads/{id}/status-history": {
      "get": {
        "summary": "リードのステータス履歴取得",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "leadId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "リードのステータス履歴",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeadStatusHistory"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/memos": {
      "get": {
        "summary": "メモ一覧取得",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "leadId": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "メモ一覧",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Memo"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "メモ作成",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "leadId": {
                    "type": "string"
                  },
                  "note": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "メモ作成",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Memo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Lead": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "organizationId": { "type": "integer" },
          "type": { "type": "string" },
          "name": { "type": "string" },
          "nameReading": { "type": "string" },
          "nickname": { "type": "string" },
          "district": { "type": "string" },
          "homePhone": { "type": "string" },
          "mobilePhone": { "type": "string" },
          "company": { "type": "string" },
          "position": { "type": "string" },
          "postalCode": { "type": "string" },
          "address": { "type": "string" },
          "phone": { "type": "string" },
          "email": { "type": "string" },
          "referrer": { "type": "string" },
          "evaluation": { "type": "integer" },
          "status": { "type": "string" },
          "statusId": { "type": "string" },
          "isPaid": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "LeadStatusHistory": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "leadId": { "type": "string" },
          "oldStatusId": { "type": "string" },
          "newStatusId": { "type": "string" },
          "organizationId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "Memo": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "leadId": { "type": "string" },
          "organizationId": { "type": "integer" },
          "type": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "note": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "NotificationPreference": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "leadId": { "type": "string" },
          "email": { "type": "boolean" },
          "line": { "type": "boolean" },
          "facebook": { "type": "boolean" },
          "instagram": { "type": "boolean" },
          "emailAddress": { "type": "string" },
          "intervalDays": { "type": "integer" },
          "organizationId": { "type": "integer" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
...
}